+++ Sources/SuperwallKit/Delegate/SuperwallDelegateObjc.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-/// The objective-c only delegate protocol that handles Superwall lifecycle events.
+/// Objective-C-only delegate protocol that handles Superwall lifecycle events.
 ///
 /// The delegate methods receive callbacks from the SDK in response to certain events that happen on the paywall.
 ///